% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - functions - lammidrive.R
\name{datesBetween}
\alias{datesBetween}
\title{List dates between two dates}
\usage{
datesBetween(date_a, date_b)
}
\description{
Function for gaining a list of all dates between two dates. Sollution to this problem by user "yifyan" at stackoverflow.com  https://stackoverflow.com/questions/14450384/create-a-vector-of-all-days-between-two-dates. date_a and date_b must be lubridate-date-objects
}
