% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - functions - lammidrive.R
\name{driveGet_temperatures}
\alias{driveGet_temperatures}
\title{get_temperatures}
\usage{
driveGet_temperatures(extend)
}
\arguments{
\item{extend}{A date. Lets you extend the temperature data into the future (assumes the coming year will be exactly like the last one)}
}
\description{
obtains the tale with the temperature logs (the manual ones) from the lammi salmon lab journal
}
\examples{
tb_temps <- lammidrive::get_temperatures().
}
