% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - functions - lammidrive.R
\name{get_mortalities}
\alias{get_mortalities}
\title{get_mortalities}
\usage{
get_mortalities(datepit_file)
}
\arguments{
\item{datepit_file}{the location of a datepit file to use for converting pit-tags to IDs}
}
\description{
Simplified function for defining the y axis with settings for tickmarks and labels.
The function is used as a ggplot object.
}
\examples{
lammidrive::get_mortalities(datepit_file="lookup - lammisalmon pit to ID - March 2021.txt").
}
