% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - functions - lammidrive.R
\name{get_tissueSamples}
\alias{get_tissueSamples}
\title{get_tissuesamples}
\usage{
get_tissueSamples()
}
\description{
obtains the table with the tissue sample data from the lammi salmon google drive
}
\examples{
tb_samples <- lammidrive::get_tissuesamples().
}
